Add new NutritionHelper class
This class will contain all of the helper methods that can be used in
the classes that implement the Nutrition Type interface.
This will mean that there will not be duplicated code in the
NutritionTypes, as there will be a single class (this helper class) that
is responsible for updating / modifying them.